removed database implementation of ProductOnCurrentDomainFacade
