+++ project-base/app/ecs.php
@@ -59,7 +59,7 @@ return static function (ECSConfig $ecsConfig): void {
             __DIR__ . '/tests/App/Functional/Model/Pricing/InputPriceRecalculationSchedulerTest.php',
             __DIR__ . '/tests/App/Smoke/BackendApiCreateProductTest.php',
             __DIR__ . '/tests/App/Smoke/Http/RouteConfigCustomization.php',
-            __DIR__ . '/tests/App/Functional/Model/Product/ProductOnCurrentDomainFacadeCountDataTest.php',
+            __DIR__ . '/tests/App/Functional/Model/Product/ProductOnCurrentDomainElasticFacadeCountDataTest.php',
             __DIR__ . '/tests/App/Functional/Model/Product/Availability/ProductAvailabilityFacadeTest.php',
             __DIR__ . '/tests/App/Functional/Model/Cart/CartMigrationFacadeTest.php',
             __DIR__ . '/src/Model/Product/Transfer/Akeneo/ProductTransferAkeneoValidator.php',
@@ -111,7 +111,7 @@ return static function (ECSConfig $ecsConfig): void {
             __DIR__ . '/src/Component/Image/ImageFacade.php',
             __DIR__ . '/src/DataFixtures/Demo/OrderDataFixture.php',
             __DIR__ . '/src/DataFixtures/Demo/ProductDataFixture.php',
-            __DIR__ . '/tests/App/Functional/Model/Product/ProductOnCurrentDomainFacadeCountDataTest.php',
+            __DIR__ . '/tests/App/Functional/Model/Product/ProductOnCurrentDomainElasticFacadeCountDataTest.php',
             __DIR__ . '/src/Model/Product/Elasticsearch/ProductExportRepository.php',
             __DIR__ . '/src/Model/Product/Product.php',
             __DIR__ . '/src/Model/Order/OrderFacade.php',
